+++ enterprise/cmd/executor-queue/main.go
@@ -9,10 +9,10 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/config"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/queues/batches"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/queues/codeintel"
-	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/server"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/batches"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/codeintel"
+	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/server"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
