+++ enterprise/cmd/frontend/executor-queue/queues/batches/queue.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"database/sql"
 
-	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/server"
+	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/server"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/background"
 	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	apiclient "github.com/sourcegraph/sourcegraph/enterprise/internal/executor"
