+++ enterprise/cmd/frontend/executor-queue/queues/batches/transform_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/config"
 	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	apiclient "github.com/sourcegraph/sourcegraph/enterprise/internal/executor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
