+++ enterprise/cmd/frontend/internal/executor/init.go
@@ -9,10 +9,10 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/config"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/queues/batches"
-	codeintel2 "github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/queues/codeintel"
-	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/executor-queue/server"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/batches"
+	codeintel2 "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/codeintel"
+	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/server"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
